Check if there's a .git directory in unit
While incorporating univ. ctags check make target with the FreeBSD port
devel/universal-ctags misc/unit is expecting if there's a git command
then you're using a git repo.

This wouldn't be the case with FreeBSD port so check to see if there's
a .git/ directory.

This isn't an absolute way to check for it but maybe a better way to
know if you should run git-ls-files.